mlcube/speech data selection/requirements.txt
pyyaml>=5.4.1
fire
protobuf
typer
gdown
sklearn
numpy
pandas
scikit-learn
pyarrow
tqdm
scipy
wget
